requirements.txt
-e .
cachecontrol==0.12.5
certifi==2018.4.16
chardet==3.0.4
django==2.1
idna==2.7
jinja2==2.10
markupsafe==1.0
msgpack==0.5.6
pyjwt==1.6.4
pytz==2018.5
requests==2.19.1
urllib3==1.23
